requirements.txt
fastapi
uvicorn
python-dotenv

# DB
sqlalchemy
psycopg2-binary
pgvector

# Redis
redis

# LLM + embeddings
openai

# File handling
pypdf

# Utils
tiktoken